Refactor USX RNC to be more RNG conformant
This changes the diagrams to something good enough but not as perfect.
The .rng still won't load in libxml due to:

grammar/usx.rng:1173: element element: Relax-NG parser error :
Element char attributes have a content type error

Which is too vague to be of value. Removing some groups from that
error fixes 3 of the 4 sites.